requirements.txt
absl-py==1.4.0
alabaster==0.7.13
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
arch==5.3.0
arm-pytorch-utilities==0.3.2
astatine==0.3.3
astor==0.8.1
astpretty==3.0.0
astroid==2.15.6
asttokens==2.4.1
attrs==23.1.0
autoflake==1.7.8
Babel==2.12.1
backcall==0.2.0
bandit==1.7.5
beautifulsoup4==4.12.3
black==23.7.0
bleach==6.1.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.4
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
cmaes==0.9.1
cmake==3.26.4
cognitive-complexity==1.3.0
colorama==0.4.6
colorlog==6.7.0
contourpy==1.1.0
coverage==7.3.0
cycler==0.11.0
Cython==0.29.21
darglint==1.8.1
DataProperty==1.0.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
dlint==0.14.1
doc8==1.1.1
docformatter==1.7.5
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.19
domdf-python-tools==3.6.1
entrypoints==0.4
eradicate==2.3.0
etils==1.5.2
exceptiongroup==1.1.3
executing==2.0.1
Farama-Notifications==0.0.4
fastapi==0.101.1
fasteners==0.19
fastjsonschema==2.19.1
filelock==3.12.2
# Fiona==1.9.4.post1
flake8==5.0.4
flake8-2020==1.8.1
flake8-aaa==0.16.0
flake8-annotations==3.0.1
flake8-annotations-complexity==0.0.8
flake8-annotations-coverage==0.0.6
flake8-bandit==4.1.1
flake8-black==0.3.6
flake8-blind-except==0.2.1
flake8-breakpoint==1.1.0
flake8-broken-line==0.6.0
flake8-bugbear==23.3.12
flake8-builtins==1.5.3
flake8-class-attributes-order==0.1.3
flake8-coding==1.3.2
flake8-cognitive-complexity==0.1.0
flake8-comments==0.1.2
flake8-comprehensions==3.14.0
flake8-debugger==4.1.2
flake8-django==1.4
flake8-docstrings==1.7.0
flake8-encodings==0.5.0.post1
flake8-eradicate==1.5.0
flake8-executable==2.1.3
flake8-expression-complexity==0.0.11
flake8-fastapi==0.7.0
flake8-fixme==1.1.1
flake8-functions==0.0.8
flake8-functions-names==0.4.0
flake8-future-annotations==0.0.5
flake8-helper==0.2.1
flake8-isort==6.0.0
flake8-literal==1.3.0
flake8-logging-format==0.9.0
flake8-markdown==0.4.0
flake8-mutable==1.2.0
flake8-no-pep420==2.7.0
flake8-noqa==1.3.2
flake8-pie==0.16.0
flake8-plugin-utils==1.3.3
flake8-pyi==22.11.0
flake8-pylint==0.2.0
flake8-pytest-style==1.7.2
flake8-quotes==3.3.2
flake8-rst-docstrings==0.3.0
flake8-secure-coding-standard==1.4.0
flake8-string-format==0.3.0
flake8-tidy-imports==4.10.0
flake8-typing-imports==1.14.0
flake8-use-fstring==1.4
flake8-use-pathlib==0.3.0
flake8-useless-assert==0.4.4
flake8-variables-names==0.0.6
flake8-warnings==0.4.0
flake8_simplify==0.20.0
fonttools==4.40.0
fsspec==2023.6.0
geopandas==0.13.2
gitdb==4.0.10
GitPython==3.1.32
glfw==2.6.4
google-auth==2.22.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.56.0
gym==0.21.0
gym-minigrid==1.0.3
# gymnasium==0.28.1
h5py==3.9.0
huggingface-hub==0.16.4
huggingface-sb3==2.2.5
hypothesis==6.82.6
hypothesmith==0.1.9
idna==3.4
imageio==2.33.1
imagesize==1.4.1
# importlib-metadata==6.8.0
importlib-resources==6.0.0
iniconfig==2.0.0
ipython==8.12.3
isort==5.12.0
jax-jumpy==1.0.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.1
jsonpickle==3.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-client==7.2.0
jupyter_core==5.7.2
jupyterlab-flake8==0.7.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.4
lark-parser==0.12.0
lazy-object-proxy==1.9.0
libcst==0.4.10
lit==16.0.6
llvmlite==0.40.1
lxml==4.9.3
Mako==1.2.4
# Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.7
mbstrdecoder==1.1.3
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
mr-proper==0.0.7
mujoco==3.1.1
mujoco-py==2.1.2.14
munch==2.5.0
mypy-extensions==1.0.0
natsort==8.4.0
nbclient==0.10.0
# nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
numba==0.57.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.76
optuna==3.2.0
packaging==24.0
panda-gym==1.1.1
Panda3D==1.10.13
panda3d-gltf==0.13
panda3d-simplepbr==0.10
pandas==1.5.3
pandas-vet==0.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.11.2
pathtools==0.1.2
pathvalidate==2.5.2
patsy==0.5.3
pbr==5.11.1
pep8-naming==0.13.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.0.0
pipreqs==0.5.0
platformdirs==3.10.0
plotly==5.15.0
pluggy==1.2.0
progressbar==2.5
prompt_toolkit==3.0.46
property-cached==1.6.4
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyaml==23.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybetter==0.4.1
pybind11==2.11.1
pybullet==3.2.5
pycln==2.2.2
pycodestyle==2.9.1
pycparser==2.21
pydantic==2.7.3
pydantic_core==2.18.4
pydocstyle==6.3.0
pyemojify==0.2.0
pyflakes==2.5.0
pygame==2.5.0
pyglet==2.0.8
Pygments==2.18.0
pylint==2.17.5
PyOpenGL==3.1.7
pyparsing==3.0.9
pyproj==3.6.0
pyproject-api==1.5.4
pytablewriter==0.64.2
pytest==7.4.0
pytest-cov==4.1.0
pytest-sugar==0.9.7
python-dateutil==2.9.0.post0
python-dev-tools==2023.3.24
pytorch-seed==0.2.0
pytz==2023.3
pyupgrade==3.10.1
PyYAML==6.0
pyzmq==19.0.2
referencing==0.35.1
removestar==1.3.1
requests==2.31.0
requests-oauthlib==1.3.1
restructuredtext-lint==1.4.0
rich==13.4.2
rliable==1.0.8
rpds-py==0.18.1
rsa==4.9
sacred==0.8.4
scikit-learn==1.3.0
scikit-optimize==0.9.0
scipy==1.11.1
seaborn==0.12.2
sentry-sdk==1.28.0
setproctitle==1.3.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
# Sphinx==6.2.1
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==2.0.18
ssort==0.11.6
stable-baselines3==1.7.0a12
stack-data==0.6.3
starlette==0.27.0
statsmodels==0.14.0
stdlib-list==0.9.0
stevedore==5.1.0
sympy==1.12
tabledata==1.3.1
tcolorpy==0.1.3
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
termcolor==2.3.0
threadpoolctl==3.1.0
tinycss2==1.3.0
tokenize-rt==5.2.0
tomli==2.0.1
tomlkit==0.12.1
torch==2.0.1
torchsummary==1.5.1
torchvision==0.15.2
tornado==6.4.1
tox==4.10.0
tox-travis==0.12
tqdm==4.65.0
traitlets==5.14.3
triton==2.0.0
typepy==1.3.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
untokenize==0.1.1
urllib3==1.26.16
virtualenv==20.26.2
wandb==0.15.5
wasabi==1.1.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.15.0
# yapf==0.40.1
yarg==0.1.9
zipp==3.16.0
